move functions to its own query-server-submissions
