+++ x/emissions/keeper/queryserver/query_server_losses.go
@@ -61,29 +61,3 @@ func (qs queryServer) GetReputerLossBundlesAtBlock(ctx context.Context, req *typ
 
 	return &types.GetReputerLossBundlesAtBlockResponse{LossBundles: reputerLossBundles}, nil
 }
-
-func (qs queryServer) GetOpenReputerSubmissionWindows(ctx context.Context, req *types.GetOpenReputerSubmissionWindowsRequest) (_ *types.GetOpenReputerSubmissionWindowsResponse, err error) {
-	defer metrics.RecordMetrics("GetOpenReputerSubmissionWindows", time.Now(), &err)
-
-	openNonces, err := qs.k.GetOpenReputerSubmissionWindows(ctx, req.TopicId)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.GetOpenReputerSubmissionWindowsResponse{
-		Nonces: &openNonces,
-	}, nil
-}
-
-func (qs queryServer) GetOpenWorkerSubmissionWindows(ctx context.Context, req *types.GetOpenWorkerSubmissionWindowsRequest) (_ *types.GetOpenWorkerSubmissionWindowsResponse, err error) {
-	defer metrics.RecordMetrics("GetOpenWorkerSubmissionWindows", time.Now(), &err)
-
-	openNonces, err := qs.k.GetOpenWorkerSubmissionWindows(ctx, req.TopicId)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.GetOpenWorkerSubmissionWindowsResponse{
-		Nonces: &openNonces,
-	}, nil
-}
